# auth-service/docs/swagger.yaml
basePath: /
definitions:
  dto.CreateUserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  dto.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  web.Error:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: rfcompanhia@hotmail.com
    name: Raphael Azevedo
    url: https://www.linkedin.com/in/raphael-a-neves/
  description: Product API with auhtentication
  termsOfService: http://swagger.io/terms/
  title: Go Auth Service
  version: "1.0"
paths:
  /user:
    post:
      consumes:
      - application/json
      description: Create
      parameters:
      - description: user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.Error'
      summary: Create
      tags:
      - users
  /user/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.Error'
      summary: Login
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"